Ticket#11 product page

+++ backend/database/migrations/2025_04_22_224436_create_products_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('Item_Name');
             $table->string('Description');
             $table->string('Category');
-            $table->decimal('Unit_Price');
+            $table->integer('Unit_Price');
             $table->integer('Quantity');
             $table->string('Image')->nullable();
             $table->timestamps();
